fixed all the errors
