+++ lib/features/healthPage/health_details_screen.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:ihealth/features/healthPage/health_video_screen.dart';
+import 'package:ihealth/common_widgets/profile_info_widget.dart.dart';
 import 'package:ihealth/models/channel_model.dart';
 import 'package:ihealth/models/video_model.dart';
 import 'package:ihealth/services/api_service.dart';
@@ -27,7 +28,7 @@ class _CourseDetailsScreenState extends State<CourseDetailsScreen> {
     });
   }
 
-  _buildVideo(Video video) {
+  Widget _buildVideo(Video video) {
     return GestureDetector(
       onTap: () => Navigator.push(
         context,
@@ -35,53 +36,25 @@ class _CourseDetailsScreenState extends State<CourseDetailsScreen> {
           builder: (_) => VideoScreen(id: video.id),
         ),
       ),
-      child: Container(
-        margin: EdgeInsets.symmetric(horizontal: 20.0, vertical: 5.0),
-        padding: EdgeInsets.all(10.0),
-        height: 140.0,
-        decoration: const BoxDecoration(
-          color: Colors.white,
-          boxShadow: [
-            BoxShadow(
-              color: Colors.black12,
-              offset: Offset(0, 1),
-              blurRadius: 6.0,
-            ),
-          ],
-        ),
-        child: Row(
-          children: <Widget>[
-            Image(
-              width: 150.0,
-              image: NetworkImage(video.thumbnailUrl),
-            ),
-            SizedBox(width: 10.0),
-            Expanded(
-              child: Text(
-                video.title,
-                style: const TextStyle(
-                  color: Colors.black,
-                  fontSize: 18.0,
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
+      child: VideoItem(video: video),
     );
   }
 
-  _loadMoreVideos() async {
+  void _loadMoreVideos() async {
     final channel = _channel;
     if (channel != null) {
-      _isLoading = true;
+      setState(() {
+        _isLoading = true;
+      });
+
       List<Video> moreVideos = await APIService.instance
           .fetchVideosFromPlaylist(playlistId: channel.uploadPlaylistId);
       List<Video> allVideos = channel.videos..addAll(moreVideos);
+      
       setState(() {
         _channel?.videos = allVideos;
+        _isLoading = false;
       });
-      _isLoading = false;
     }
   }
 
@@ -124,84 +97,52 @@ class _CourseDetailsScreenState extends State<CourseDetailsScreen> {
               child: CircularProgressIndicator(
                 valueColor: AlwaysStoppedAnimation<Color>(
                   Theme.of(context).primaryColor,
-                ),
+                ), 
               ),
             ),
     );
   }
 }
 
-class BuildProfileInfo extends StatelessWidget {
-  final Channel? channelParam;
+class VideoItem extends StatelessWidget {
+  final Video video;
 
-  const BuildProfileInfo({
-    Key? key,
-    this.channelParam,
-  }) : super(key: key);
+  const VideoItem({required this.video});
 
   @override
   Widget build(BuildContext context) {
-    final channel = channelParam;
-
-    return channel != null
-        ? Container(
-            margin: EdgeInsets.all(25.0),
-            padding: EdgeInsets.all(20.0),
-            height: 100.0,
-            decoration: const BoxDecoration(
-              color: Colors.white,
-              boxShadow: [
-                BoxShadow(
-                  color: Colors.black12,
-                  offset: Offset(0, 1),
-                  blurRadius: 6.0,
-                ),
-              ],
+    return Container(
+      margin: EdgeInsets.symmetric(horizontal: 20.0, vertical: 5.0),
+      padding: EdgeInsets.all(10.0),
+      height: 140.0,
+      decoration: const BoxDecoration(
+        color: Colors.white,
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black12,
+            offset: Offset(0, 1),
+            blurRadius: 6.0,
+          ),
+        ],
+      ),
+      child: Row(
+        children: <Widget>[
+          Image(
+            width: 150.0,
+            image: NetworkImage(video.thumbnailUrl),
+          ),
+          SizedBox(width: 10.0),
+          Expanded(
+            child: Text(
+              video.title,
+              style: const TextStyle(
+                color: Colors.black,
+                fontSize: 18.0,
+              ),
             ),
-            child: Row(
-              children: [
-                SizedBox(width: 12.0),
-                CircleAvatar(
-                  backgroundColor: Colors.white,
-                  radius: 35.0,
-                  backgroundImage: NetworkImage(channel.profilePictureUrl),
-                ),
-                SizedBox(width: 12.0),
-                Expanded(
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: <Widget>[
-                      Text(
-                        channel.title,
-                        style: const TextStyle(
-                          color: Colors.black,
-                          fontSize: 20.0,
-                          fontWeight: FontWeight.w600,
-                        ),
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                      Text(
-                        '${channel.subscriberCount} subscribers',
-                        style: TextStyle(
-                          color: Colors.grey[600],
-                          fontSize: 16.0,
-                          fontWeight: FontWeight.w600,
-                        ),
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          )
-        : const SizedBox();
+          ),
+        ],
+      ),
+    );
   }
-}
-
-void main() {
-  runApp(MaterialApp(
-    home: CourseDetailsScreen(),
-  ));
 }
