+++ lib/models/course_model.dart
@@ -1,0 +1,7 @@
+class Course {
+  final String title;
+  final int progress;
+  final String imageUrl;
+
+  Course(this.title, this.progress, this.imageUrl);
+}
