+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   http: ^1.2.0
   youtube_player_flutter: ^8.1.2
+  provider: ^6.1.1
 
 dev_dependencies:
   flutter_test:
